Dropbox module and API updates
Dropbox - started a module for dropbox support.  Wraps the dropbox
python API and creates support for easily writing apps for dropbox
integration.  Handles the app key and secret key and takes care of the
request tokens for authentication
API - Path objects have functions that wouldn't work on linux because of
case sensitivity, this should hopefully be fixed but needs more
multiplatform testing
API - found a great progress bar class for the command line, would like
to have a suite of utils for command line tools eventually