New Widget - ProgressWidget
GUI - Added a new ProgressWidget.  It works pretty well but i would also
like to support multiple progressbars within the same widget.
GUI - Loginwidget trys to auto login if the credentials are read from a
file
GUI - Path widget got a bug fix to the right click menu "Open in
Explorer" command
DB - This module is so outdated right now, tried to add a progress data
for supporting scenegraph style progresses for multiprogressbar widget
but DB needs an overhaul to get back up to speed.
API - Logger tracks if a streamlogger was setup so multiple can't
happen.  I'm not thrilled with the logger class right now, but its not
terriable so i'll fix it at some later date.