Progress widget update
GUI - So i decided that the progress widget shouldn't support multiple
because its no really feasible in top down code evaluation.  This is
best for a specialized class that can get the full setup prior to
running, build the progress tree, and then somehow execute through it
all.  I've forgone this because its complex and i'd rather get the DB
stuff in working order again.