Fixing up logging
DTL - __pkgname__ will change based on where you are running, but now it
should just find where DTL is and used its directory name as the
__pkgname__.  This affects logging because we create a log and db log
files from this name.  I'm still not happy with how __init__.py houses
the variables.  There should be a config file somewhere that takes care
of most of this, especially the log setup, and this should only contain
actual internal information not customizable information